fixed some issues with serialization of nested queries into WHERE
When removing the nesting operator and its result attribute from downstream
operators, this changes the schemas of downstream operators. Some of these
operators may be LHS inputs to nesting operators, i.e., their attributes can be
referenced in correlations. The updated code now creates global data structures
for nested subqueries during serialization. This includes bookkeeping a list of
correlated attribute references for each nested subquery such that we can adjust
the attribute positions when the schema of the referenced outer relation
changes.

Currently, we complain but do not fail if during updating of attribute names for
referencing FROM items in the translated SQL query we recognize that attribute
names have chagned. This is necessary as join may rename result attributes and
this is expected. Still open:

1) in some cases the FromAttrsContext is not adjusted yet (still has nesting
result attributes) when it is first used and also for double nesting the outer
scopes we push seem wrong in some cases.